Fix for 0-byte malloc error (Electric Fence)
This was a problem(?) in the GNU regex library caught by Electric Fence.
Thus, we apply this workaround to avoid the error. Commit 48af disabled linking
with Electric Fence to avoid this error. This commit specifically addresses
the zero-byte allocation error instead of disabline Electric Fence altogether.
NOTE: This is reworked fix to replace df4fb.